Scripts: Add batchelor.
Add a new python package called batchelor, which will become a python wrapper
library around the various batch systems in use so that one can write scripts
using batchelor and if one changes batch systems, only a config file has to be
adjusted for everything to work again. Barely any code is committed here, but
work on this will continue.