Scripts: Add first actual function to batchelor.
Add code to check the number of running jobs with a certain job name on gridka.
Also add some skeleton code for the same task for the other batch systems,
which will just raise a "not implemented" exception if called. Finally, revise
the runCommand function so that it is easier to handle and fix a bug which
prevented stderr to be returned correctly.