Scripts: Improve batchelor interface.
Improve the interface of the batchelor by modifying the functions so one can
obtain a list of the active jobs. This prevents the case that one gets the
number of active jobs and then has to poll each job id which one had saved to
see which of the jobs had finished.