Scripts: Fix handling of many jobs in batchelor.
After the first try to fix the problem, getListOfActiveJobs still crashed if
there where many active jobs. As it turned out, the problem was not with the
limited subprocess buffer, but with the output from qstat. If there are too
many jobs, qstat closes the xml output and starts again with a new xml header
for some reason, which the ElementTree from python could not handle. Split the
output from qstat now at every xml header and parse the resulting files one by
one, which should hopefully finally fix this.