Scripts: Add C2PAP to batchelor.
Add the C2PAP system to the available clusters in the batchelor. No actual code
for interacting is there yet, just the decoration. Also fix some minor things
in the executable scripts. They should probably be deleted at some point, as
they were only thought to be for quick and dirty testing.