Add simulator submodule.
To test code which uses batchelor without actually submitting any jobs to any
batch system, add a simulator module which provides the same commands and
interface and will keep track of submitted jobs and so on without actually
doing anything. Jobs submitted to the simulator have a life time which can be
set in the config file for which they are shown as running and after which they
are deleted.