Add function to submit multiple jobs at once
Add a function to be able to submit several jobs with one call to
batchelor. The argument to the new 'submitJobs' function is a list of
the arguments of the 'submitJob' function. In case an implementation
does not provide a 'submitJobs' function itself, 'submitJob' is called
for each job to be submitted.

The return value of 'submitJobs' is a list of the job IDs. An ID of -1
means that there was an error while submitting. Unfortunately the
exceptions cannot be passed through if 'submitJobs' is not implemented.