Clean shutdown for Local option
If a SIGINT (e.g., CTRL+C) was send to the main process, it was also
send to all its child processes. This could be ignored for the worker
processes, but not for the Manager process (at least with python 2.6).
With a missing Manager process, also the worker processes stoppped.
Using threads instead of processes seems to cure this problem.