Fixed submitJob:
It was not possible to submit array jobs with the main submitJob function,
because the variables for the array were missing. I added them now and
it is now possible to submit jobs at least for the e18/lxplus system also
as array jobs.
For C2PAP I also added the variables to the funcion, but submitting array
jobs is NOT IMPLEMENTED yet.