Fixed the pull request.
For the submitArrayJob function the default array step is now set to 1. I
also change the "One" into "one" in the antok exception string. Spaces
are added to the inbuilt range() function in the GridKA section.
Comments with an example output for the return value of the
getListOfActiveJobs() are added in each batch system section.
The getListOfActive()function returns now always the same type for each
bacht system: [(jobId, taskId, jobStatus), ...]
For the C2PAP batch system the output is filled with dummy values for
the taskId and the jobStatus like: [(jobId, -1, "dummy"), ...]