Fix issues with zero length job list in 'local'
Two similar issues with a zero length job list for the 'local' system
were present in the 'deleteJobs' function, and in the way the workers
obtained the next job. In both cases the variable used to loop over
the jobs would never be initialized as there was nothing to loop
over, but would still be used in the following code. Now create a
list of the job IDs and operate on this list.