Support for Long queue on E18 system
The support for setting the queue directly for the E18 batch system
is implemented. The parameter "shortqueue" is changed to "queue" and
the values are now "short" or "Short", "medium" or "Medium", "long"
or "Long".
The default is still the short queue.
This change is breaking! Since the parameter name is changed from
"shortqueue" to "queue" - results into a initialization error in
batchelor.