BatchelorHandler: Implement two lits of job ids to handle collected jobs
One list `_submittedJobs` holds to job ids of actually submitted jobs.
For collected jobs, a -1 is entred into the list until the collected
jobs are submitted.

One list `_jobIds` for internal identification of individual jobs.
If we do not collect jobs, `_jobIds` is equal to `_submittedJobs`.
If we collect jobs, `_jobIds` contains an internal id for each job (each
call to `submitJob`).
This list is used for internal referencing the individual jobs (not the
jobs on the batch system).